pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claudette"
version = "0.1.0"
description = "A powerful CLI chatbot powered by Ollama with advanced tool support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["ollama", "chatbot", "cli", "ai", "llm", "assistant"]
authors = [
    {name = "W4pity", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Dependencies extraites de requirements.txt (sans les versions épinglées pour la flexibilité)
dependencies = [
    "ollama>=0.6.0",
    "rich>=14.0.0",
    "prompt_toolkit>=3.0.0",
    "PyYAML>=6.0.0",
    "tiktoken>=0.12.0",
    "colorama>=0.4.6",
    "ddgs>=0.1.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "fake-useragent>=2.0.0",
]

[project.optional-dependencies]
# Dépendances de développement
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

# Dépendances pour créer des binaires
build = [
    "pyinstaller>=5.0.0",
]

# Toutes les dépendances optionnelles
all = [
    "claudette[dev,build]",
]

[project.urls]
Homepage = "https://github.com/w4pity/claudette"
Documentation = "https://github.com/w4pity/claudette#readme"
Repository = "https://github.com/w4pity/claudette.git"
"Bug Tracker" = "https://github.com/w4pity/claudette/issues"

# Point d'entrée de l'application
[project.scripts]
claudette = "claudette.main:main"

# Configuration de setuptools
[tool.setuptools]
packages = ["claudette", "claudette.commands", "claudette.tools_impl", "claudette.utils"]

[tool.setuptools.package-dir]
claudette = "src"

[tool.setuptools.package-data]
claudette = ["py.typed"]

# Configuration Black (formattage du code)
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Configuration isort (tri des imports)
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_gitignore = true

# Configuration pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--color=yes",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# Configuration mypy (vérification de types)
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
